+++ src/state/droppable-dimension.js
@@ -166,12 +166,12 @@ export const scrollDroppable = (
   return result;
 };
 
-// TODO
-const growSubjectIfNeeded = ({
-  draggables: DraggableDimensionMap,
-  droppable: DroppableDimension,
-  addition: Position,
-}): DroppableDimension => {
+// TODO: make this work
+// const growSubjectIfNeeded = ({
+//   draggables: DraggableDimensionMap,
+//   droppable: DroppableDimension,
+//   addition: Position,
+// }): DroppableDimension => {
 
-};
+// };
